Add age secret management and move some services
I've had secrets in my config for some time but did not want to push
them before considering the implications of having secrets in git.
age encryption is probably good enough

I also moved some services from the common to the graphical bit that are
probably not very useful on a server.