Update flake inputs and add cosmic
Note that the update required switching out the code base for the window
button applet, which would no longer work with the patch from the
original repo.

Also, cosmic is currently disabled since (even with the substitutor) it
sometimes tries to rebuild from source.
It seems like I would have to switch my nixpkgs to the one from the
flake, which I don't want to do.
May take a look at that again once the beta is here...